refactor(core): address PR comments, clean up service, and remove redundant memory bounds
