+++ packages/core/src/agents/local-executor.ts
@@ -1033,7 +1033,6 @@ export class LocalAgentExecutor<TOutput extends z.ZodTypeAny> {
         startHistory,
         undefined,
         undefined,
-        'subagent',
       );
     } catch (e: unknown) {
       await reportError(
